moved files and begin backend server setup
